refactor(checker): extract infer-binding helpers into dedicated module
Move the five infer-binding collection methods from the oversized
queries/class.rs (was 2069 lines, over the 2000-line guard) into a
new queries/infer_bindings.rs, bringing class.rs to 1578 lines.

No behavior change — the public interfaces and call sites are unchanged.
